Do not start a deployment if initialize is enabled
This adds the ability to create a one-time initialize job which starts
a job with a app pod to run an `initalizeCommand`. Because the
deployment will be useless until initialization has been run, the
app deployment will *NOT* be rendered/loaded.

A second release where the deployment will be rendered/loaed is thus
needed.

Moves all db hook related values to be under `application` too.

Example :

```
  # first release
  helm install \
    --set application.initializeCommand="bin/setup"

  # second release
  helm upgrade \
    --reuse-values
    --set application.initializeCommand=""
```