Allow a Secret to be used in application container
This creates an option where users can pass in a setting to specify the
name of the Secret. This name will be used for deployments to load the
Secret's key-value pairs as environment variables in the application
container.

This option is optional. The Secret must be created by the user
beforehand.